Move spectral models to gammapy/modeling/models
