+++ gammapy/modeling/models/spectral.py
