+++ gammapy/modeling/models/spectral_cosmic_ray.py
@@ -5,7 +5,7 @@ For measurements, the "Database of Charged Cosmic Rays (CRDB)" is a great resour
 http://lpsc.in2p3.fr/cosmic-rays-db/
 """
 from astropy import units as u
-from .core import PowerLaw, SpectralLogGaussian
+from .spectral import PowerLaw, SpectralLogGaussian
 
 __all__ = ["create_cosmic_ray_spectral_model"]
