+++ gammapy/modeling/models/spectral_crab.py
@@ -1,7 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import numpy as np
 from astropy import units as u
-from .core import ExponentialCutoffPowerLaw, LogParabola, PowerLaw, SpectralModel
+from .spectral import (
+    ExponentialCutoffPowerLaw,
+    LogParabola,
+    PowerLaw,
+    SpectralModel,
+)
 
 __all__ = ["create_crab_spectral_model"]
