+++ gammapy/modeling/models/tests/test_spectral.py
