+++ gammapy/modeling/models/tests/test_spectral_crab.py
