+++ gammapy/utils/serialization/tests/test_xml.py
@@ -2,8 +2,7 @@
 import pytest
 import numpy as np
 from numpy.testing import assert_allclose
-from gammapy.modeling.models import SkyModels, spatial
-from gammapy.modeling.models import spectrum as spectral
+from gammapy.modeling.models import SkyModels, spatial, spectral
 from gammapy.utils.serialization import UnknownModelError, xml_to_sky_models
 from gammapy.utils.testing import requires_data
